Fix relative target of manpage links.
Whilst working on the Reproducible Builds effort [0] we noticed that mocassin
could not be built reproducibly.

This led me to discover that the mocassinOutput(1), mocassinPlot(1) and
mocassinWarm(1) manpages were linking to the absolute location for the
mocassin(1) manpage, so they would not work when distributed:

For example mocassinOutput.1.gz linked to, on my machine:

  /home/lamby/temp/cdt.20210215121553.ZpEolGOptg.repro.mocassin/build-a/mocassin-2.02.73.1/debian/tmp/usr/share/man/man1/mocassin.1.gz

Patch attached that uses relative URLs instead. This change would also make the
build reproducible. I originally filed this in Debian as #982851 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/982851